move otel setup in to separate file
